refactor: enhance title block processing and improve markdown generation
